+++ lib/rules/require-data-selectors.js
